Fix failing behave tests
